% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{cumrate}
\alias{cumrate}
\title{Calculate the cumulative rate.}
\usage{
cumrate(
  count,
  pop,
  rate = NULL,
  eage = 70,
  agewidth = 5,
  sep_zero = TRUE,
  mp = 1,
  decimal = 5
)
}
\arguments{
\item{count}{Number of cases number.}

\item{pop}{Number of population at risk.}

\item{rate}{Age specific incidence or mortality rate.}

\item{eage}{The defined age limit of the calculated cumulative rate.}

\item{agewidth}{The age band width of the age specific rate.}

\item{sep_zero}{Logical value, if the 0 age group was separate.}

\item{mp}{A multiplier used to scale the calculated rates. Default is 100.}

\item{decimal}{Decimals of the calculated rates.}
}
\value{
Cumulative rate.
}
\description{
It was used to calculate the cumulative rate up to a defined age limit.
}
\examples{
px <- c(20005, 86920, 102502, 151494, 182932, 203107, 240289, 247076, 199665,
        163820, 145382, 86789, 69368, 51207, 39112, 20509, 12301, 6586, 1909)
dx <- c(156, 58, 47, 49, 48, 68, 120, 162, 160, 294, 417, 522, 546, 628,
        891, 831, 926, 731, 269)
mx <- dx / px
cumrate(mx, eage = 70)
}
