% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr.R
\name{estimate.list}
\alias{estimate.list}
\title{Estimate the lifetime risk and 95\% confidence interval}
\usage{
\method{estimate}{list}(x, sage = 0, mp = 100, decimal = 2)
}
\arguments{
\item{x}{List that contains elements with class of 'ltr'.}

\item{sage}{The initial age of the lifetime risk.}

\item{mp}{Multiple}

\item{decimal}{Number of decimal}
}
\value{
Life time risk and 95\% CI.
}
\description{
Estimate the lifetime risk and 95\% confidence interval
}
