% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr.R
\name{estimate.ltr}
\alias{estimate.ltr}
\title{Estimate the lifetime risk and 95\% confidence interval}
\usage{
estimate.ltr(x, sage = 0)
}
\arguments{
\item{x}{Object of class 'ltr' outputed by ltr function.}

\item{sage}{The initial age of the lifetime risk.}
}
\value{
Life time risk and 95\% CI.
}
\description{
Estimate the lifetime risk and 95\% confidence interval
}
