% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr.R
\name{estimate}
\alias{estimate}
\title{Estimate the lifetime risk and 95\% confidence interval}
\usage{
estimate(x, sage = 0, mp = 100, decimal = 2)
}
\arguments{
\item{x}{Object of class 'ltr' outputed by ltr function.}

\item{sage}{The initial age of the lifetime risk.}

\item{mp}{Multiple}

\item{decimal}{Number of decimal}
}
\value{
Life time risk and 95\% CI.
}
\description{
Estimate the lifetime risk and 95\% confidence interval
}
