% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk.R
\name{get_risk}
\alias{get_risk}
\alias{get_risk.devcan}
\alias{get_risk.amp}
\alias{get_risk.wun}
\alias{get_risk.cumu}
\title{Calculate the risk value for a specific age range}
\usage{
get_risk(object, age_start, age_end, age_combine = c("cross", "pairwise"))

\method{get_risk}{devcan}(object, age_start, age_end, age_combine = c("cross", "pairwise"))

\method{get_risk}{amp}(object, age_start, age_end, age_combine = c("cross", "pairwise"))

\method{get_risk}{wun}(object, age_start, age_end, age_combine = c("cross", "pairwise"))

\method{get_risk}{cumu}(object, age_start, age_end, age_combine = c("cross", "pairwise"))
}
\arguments{
\item{object}{Returned object from a risk model calculated using the devcan,
amp, or wun function.}

\item{age_start}{Starting age(s) used to calculate risk over one or
multiple age ranges.}

\item{age_end}{Ending age(s) used to calculate risk over one or
multiple age ranges.}

\item{age_combine}{How age vectors are combined: \code{"cross"} or \code{"pairwise"}.}
}
\value{
Risks according to the age ranges.
}
\description{
Calculate the risk value for a specific age range
}
