% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr.R
\name{ltr}
\alias{ltr}
\title{Estimate the age conditional probability of developing or dying from cancer}
\usage{
ltr(mi, di, ri, ni, type = "developing", age_width = 5)
}
\arguments{
\item{mi}{The annual number of all-cause mortality deaths in each age group.}

\item{di}{The annual number of cancer-related deaths in each age group.}

\item{ri}{The annual number of diagnosed cancer cases in each age group.}

\item{ni}{The number of population in each age group.}

\item{type}{Characters "developing" or "dying" indicate estimate the
probability of developing cancer or dying from it.}

\item{age_width}{The age width of each age group.}
}
\value{
A list with class of ltr:
\item{age}{Age groups.}
\item{si}{The probability of developing or dying from cancer
in each age group.}
\item{vari}{The variance of the estimates for each age group.}
}
\description{
Estimate the age conditional probability of developing or dying from cancer
}
\examples{
ni <- c(
  73872987, 82029530, 72267070, 78303514, 99425613, 119915673, 98068725,
  96644427, 121225951, 121250720, 96012917, 79863455, 75972753, 52929797,
  37551107, 29047207, 19584254, 13854299
)
mi <- c(
  60594, 17718, 18883, 28127, 37493, 75223, 83574, 100655, 211467, 278913,
  419663, 445223, 770865, 929008, 1058922, 1346942, 1576852, 2305312
)
di <- c(
  3511, 2801, 2553, 3183, 4960, 9456, 13509, 23935, 62386, 111640, 147866,
  203955, 301892, 304985, 302785, 323804, 275557, 197614
)
ri <- c(
  9303, 6887, 6248, 8509, 16961, 39439, 56670, 86535, 189251, 289320, 344395,
  411232, 552071, 491213, 433786, 395544, 292672, 173503
)
ll <- ltr(mi, di, ri, ni)
}
