% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr.R
\name{ztest}
\alias{ztest}
\title{Test the difference of lifetime risk between two groups using Z test}
\usage{
ztest(x, y)
}
\arguments{
\item{x}{Age conditional probability in group x with class of 'ltr'
returned by function of ltr.}

\item{y}{Age conditional probability in group y with class of 'ltr'
returned by the ltr function.}
}
\value{
The statistics Z and P of the test result.
}
\description{
Test the difference of lifetime risk between two groups using Z test
}
